+++ mypyc/codegen/emitclass.py
